% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAPI.R
\name{generateURL}
\alias{generateURL}
\title{Generate a URL for the Disqus API call.}
\usage{
generateURL(
  ressource = c("forums", "pages", "posts", "threads", "trends", "users"),
  option = NULL,
  forum = NULL,
  thread = NULL,
  user = NULL,
  post = NULL,
  author = NULL,
  category = NULL,
  count = NULL,
  depth = NULL,
  interval = NULL,
  query = NULL,
  range = NULL,
  since = NULL,
  since_id = NULL,
  start = NULL,
  end = NULL,
  priority = NULL,
  attach = NULL,
  filters = NULL,
  include = NULL,
  related = NULL,
  limit = 25,
  order = "desc",
  cursor = NULL,
  type = "json",
  version = "3.0",
  key = NULL
)
}
\arguments{
\item{ressource}{Character vector specifying the type of resource to access through the API.}

\item{option}{Character vector specifying the specific API operation to perform.}

\item{forum}{Character vector containing the forum identifier or shortname.}

\item{thread}{Character vector containing the thread identifier or link.}

\item{user}{Character vector containing the user identifier.}

\item{post}{Character vector containing the post identifier.}

\item{author}{Character vector containing the author identifier.}

\item{category}{Character vector containing the category identifier.}

\item{count}{Numeric value specifying the number of items to retrieve.}

\item{depth}{Numeric value specifying the depth of the query.}

\item{interval}{Character vector specifying the time interval for data retrieval.}

\item{query}{Character vector containing the query string for filtering data.}

\item{range}{Character vector specifying the time range for data retrieval.}

\item{since}{Character vector specifying the starting date for data retrieval.}

\item{since_id}{Character vector specifying the starting identifier for data retrieval.}

\item{start}{Character vector specifying the starting point for data retrieval.}

\item{end}{Character vector specifying the ending point for data retrieval.}

\item{priority}{Character vector specifying the sorting type for data retrieval.}

\item{attach}{Character vector specifying additional attachments for the query.}

\item{filters}{Character vector specifying filters to apply to the query.}

\item{include}{Character vector specifying additional data to include in the response.}

\item{related}{Character vector specifying related data to include in the response.}

\item{limit}{Numeric value specifying the maximum number of items to return in the response.}

\item{order}{Character vector specifying the order of the items in the response.}

\item{cursor}{Character vector specifying the cursor for pagination.}

\item{type}{Character vector specifying the data type for the response (e.g., "json", "xml").}

\item{version}{Character vector specifying the version of the Disqus API to use (default is "3.0").}

\item{key}{Character vector containing the API key. If not provided, a placeholder is used.}
}
\value{
A character vector representing the API call URL, based on the provided parameters.
}
\description{
This function generates the URL for making API calls to the Disqus platform.
It constructs the URL based on the specified resource, option, and various parameters provided.
The generated URL can then be used to retrieve data from the Disqus API.
}
\examples{
# Generate the URL to retrieve 10 posts from the "my_forum" forum in JSON format
generateURL(ressource = "posts", option = "list", forum = "my_forum", thread = "XYZ")

}
