% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_thread_list}
\alias{get_thread_list}
\title{Get a list of threads from the Disqus API.}
\usage{
get_thread_list(
  thread,
  forum = NULL,
  key,
  second_key = NULL,
  verbose = TRUE,
  n = 100,
  limit = 25,
  format = c("tidy", "raw"),
  ...
)
}
\arguments{
\item{thread}{Character vector containing the thread identifier.}

\item{forum}{Character vector containing the forum identifier or shortname (default is NULL).}

\item{key}{Character vector containing the primary API key to authenticate the API call(s).}

\item{second_key}{Character vector containing a second API key to use if the hourly API limit is reached with the primary key (default is NULL).}

\item{verbose}{Logical value indicating whether to display informative messages during the API calls (default is TRUE).}

\item{n}{The maximum number of observations to retrieve (default is 100).}

\item{limit}{Numeric value specifying the maximum number of items to return in each API response (default is 25).}

\item{format}{Character vector specifying the desired output format: 'tidy' (unnested) or 'raw' (JSON) (default is 'tidy').}

\item{...}{Additional arguments to be passed to the \code{generateURL} function for URL generation.}
}
\value{
The retrieved list of threads from the Disqus API in the specified format.
}
\description{
This function retrieves a list of threads from the Disqus API for the specified thread identifier and forum (optional).
The data is returned in either a 'tidy', unnested format or 'raw' (JSON) format.
}
\examples{
# Retrieve a list of threads in 'tidy' format using the primary key
# tidy_threads <- get_thread_list(thread = "thread_id", key = "YOURAPPKEY")

}
\seealso{
\code{generateURL}, \code{paginateAPI}, \code{get_generic}, \code{unnest_recursively}
}
