% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAPI.R
\name{leechAPI}
\alias{leechAPI}
\title{Call the Disqus API using spare key.}
\usage{
leechAPI(url, key, second_key = NULL, verbose = TRUE)
}
\arguments{
\item{url}{A character vector representing the complete URL for the Disqus API call.}

\item{key}{A character vector containing the primary API key to authenticate the first API call.}

\item{second_key}{A character vector containing a second API key to use if the hourly API limit is reached with the primary key (default is NULL).}

\item{verbose}{Logical value indicating whether to display informative messages (default is TRUE).}
}
\value{
The response from the Disqus API call, parsed as a list.
}
\description{
This function makes a call to the Disqus API using the provided URL and primary API key.
If the hourly API limit is reached (error code 13), and a second API key is provided,
the function will make a second API call using the second key.
The response from the API call is parsed and returned as a list.
}
\examples{
# Call the API to retrieve the data from the specified URL using the primary key,
# with a fallback to the second key if the hourly API limit is reached.
#result <- leechAPI(url = url,
#                   key = "YOURAPPKEY",
#                   second_key = "YOURSECONDAPPKEY")

}
