% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callAPI.R
\name{paginateAPI}
\alias{paginateAPI}
\title{Call the Disqus API using pagination and optionally a second API key.}
\usage{
paginateAPI(url, key, second_key = NULL, verbose = TRUE, n = Inf)
}
\arguments{
\item{url}{A character vector representing the complete URL for the initial Disqus API call.}

\item{key}{A character vector containing the primary API key to authenticate the first API call.}

\item{second_key}{A character vector containing a second API key to use if the hourly API limit is reached with the primary key (default is NULL).}

\item{verbose}{Logical value indicating whether to display informative messages during API calls (default is TRUE).}

\item{n}{Numeric value specifying the maximum number of observations to retrieve (default is Inf, meaning all available observations).}
}
\value{
A list or vector containing the combined data retrieved from multiple API calls, based on pagination and API keys.
}
\description{
This function makes API calls to the Disqus API using the provided URL and primary API key. If a second API key is provided, and the hourly API limit is reached with the primary key, the function will use the second key for subsequent API calls. The function implements pagination to retrieve multiple pages of data if available and combines the results into a single R object.
}
\examples{
# Call the API to retrieve all available data from the specified URL using the primary key,
# with a fallback to the second key if the hourly API limit is reached.
#all_data <- paginateAPI(url = url,
#                         key = "YOURAPPKEY",
#                         second_key = "YOURSECONDAPPKEY",
#                         verbose = TRUE,
#                         n = Inf)

}
